[feat]部分的なパーサを公開
